chore: move App component from root folder to components folder
